+++ packages/cubejs-jdbc-driver/src/JDBCDriver.js
@@ -12,18 +12,25 @@ const DriverManager = require('jdbc/lib/drivermanager');
 const Connection = require('jdbc/lib/connection');
 const DatabaseMetaData = require('jdbc/lib/databasemetadata');
 const jinst = require('jdbc/lib/jinst');
-const mvn = promisify(require('node-java-maven'));
+const mvn = require('node-java-maven');
 
 let mvnPromise = null;
 
 const initMvn = (customClassPath) => {
   if (!mvnPromise) {
-    mvnPromise = mvn().then((mvnResults) => {
-      if (!jinst.isJvmCreated()) {
-        jinst.addOption('-Xrs');
-        const classPath = mvnResults.classpath.concat(customClassPath || []);
-        jinst.setupClasspath(classPath);
-      }
+    mvnPromise = new Promise((resolve, reject) => {
+      mvn((err, mvnResults) => {
+        if (err && !err.message.includes('Could not find java property')) {
+          reject(err);
+        } else {
+          if (!jinst.isJvmCreated()) {
+            jinst.addOption('-Xrs');
+            const classPath = (mvnResults && mvnResults.classpath || []).concat(customClassPath || []);
+            jinst.setupClasspath(classPath);
+          }
+          resolve();
+        }
+      });
     });
   }
   return mvnPromise;