+++ rust/cubesql/cubesql/src/compile/rewrite/mod.rs
@@ -199,6 +199,9 @@ crate::plan_to_language! {
             op: String,
             values: Vec<String>,
         },
+        SegmentMember {
+            member: String,
+        },
         FilterOp {
             filters: Vec<LogicalPlan>,
             op: String,
@@ -549,6 +552,10 @@ fn filter_op_filters(left: impl Display, right: impl Display) -> String {
 
 fn filter_member(member: impl Display, op: impl Display, values: impl Display) -> String {
     format!("(FilterMember {} {} {})", member, op, values)
+}
+
+fn segment_member(member: impl Display) -> String {
+    format!("(SegmentMember {})", member)
 }
 
 fn measure_expr(measure_name: impl Display, expr: impl Display) -> String {